Move cli_parser, common & default_runtime macros to top levele module
