+++ module-system/sov-modules-macros/src/dispatch/dispatch_call.rs
@@ -1,8 +1,6 @@
-use super::common::get_serialization_attrs;
-use super::common::parse_generic_params;
-use super::common::StructDef;
-use super::common::StructFieldExtractor;
-use super::common::CALL;
+use crate::common::{
+    get_serialization_attrs, parse_generic_params, StructDef, StructFieldExtractor, CALL,
+};
 use syn::DeriveInput;
 
 impl<'a> StructDef<'a> {
