+++ module-system/sov-modules-macros/src/dispatch/mod.rs
@@ -1,6 +1,3 @@
-pub mod cli_parser;
-pub(crate) mod common;
-pub mod default_runtime;
 pub mod dispatch_call;
 pub mod genesis;
 pub mod message_codec;
