+++ module-system/sov-modules-macros/src/module_info.rs
@@ -1,4 +1,4 @@
-use super::dispatch::common::parse_generic_params;
+use crate::common::parse_generic_params;
 use proc_macro2::{self};
 use proc_macro2::{Ident, Span};
 use syn::{DataStruct, DeriveInput, ImplGenerics, PathArguments, TypeGenerics, WhereClause};
